👩🏾‍💻 Rename from modal to AddColumnModal
